fix(xhs-cloud): cast is_virtual for scan_delta COALESCE
monitor_goods.is_virtual is boolean while premium_goods.is_virtual is integer; cast to int before COALESCE in fetch_items_from_scan_delta.

+++ xhs-cloud/cloud_deploy/reporting/pg_reader.py
@@ -641,7 +641,7 @@ def fetch_items_from_scan_delta(
                    sd.goods_id, sd.sold_num, sd.delta, sd.deal_price,
                    prev.sold_num AS prev_sold,
                    COALESCE(m.title, pg.title) AS title,
-                   COALESCE(m.is_virtual, pg.is_virtual) AS is_virtual,
+                     COALESCE(m.is_virtual::int, pg.is_virtual) AS is_virtual,
                    COALESCE(m.pool, 'WATCH') AS pool,
                    COALESCE(m.store_id, pg.store_id) AS store_id,
                    COALESCE(m.store_name, pg.store_name) AS store_name,
